# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AJVC Portfolio Intelligence Dashboard API",
    "description": "\n## 🚀 AJVC Portfolio Intelligence Dashboard - Backend API\n\nA comprehensive **portfolio management system** for venture capital firms to track portfolio companies,\nanalyze project execution, monitor commitments, and gain AI-powered insights from meeting notes.\n\n### Key Features\n\n* **📊 Company Management**: Track portfolio companies with investment details and status\n* **📈 Gantt Layer Analytics**: Pull data from Google Sheets and compute execution KPIs\n* **📝 AI-Powered Meeting Analysis**: Upload DOCX meeting notes for automatic parsing with Google Gemini\n* **✅ Commitment Tracking**: Automatically extract and track actionable commitments from meetings\n* **🔒 Secure Authentication**: JWT-based authentication for all protected endpoints\n\n### Tech Stack\n\n* **FastAPI** for high-performance async API\n* **PostgreSQL** with SQLAlchemy ORM\n* **Google Gemini AI** for intelligent meeting parsing\n* **Google Sheets API** for Gantt data integration\n* **JWT Authentication** for secure access\n\n### Getting Started\n\n1. Authenticate via `/api/v1/auth/login` to get a JWT token\n2. Include the token in the `Authorization: Bearer <token>` header\n3. Create or select a company via `/api/v1/companies`\n4. Upload meeting notes or pull Gantt data for analysis\n\n### API Versioning\n\nCurrent version: **v1.0.0** - All endpoints are prefixed with `/api/v1`\n    ",
...
  "paths": {
...
    "/api/v1/companies/{company_id}/meetings/upload-docx": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Upload Meeting Docx",
        "description": "Upload a DOCX file containing ONE meeting's notes.\nParses the meeting and stores in database.",
        "operationId": "upload_meeting_docx_api_v1_companies__company_id__meetings_upload_docx_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_meeting_docx_api_v1_companies__company_id__meetings_upload_docx_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/meetings/upload-test-dump": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Upload Test Meeting Dump",
        "description": "TEST ENDPOINT: Upload Meeting_Dump.docx with multiple meetings.\nThis is for testing only. Production uses single meeting uploads.",
        "operationId": "upload_test_meeting_dump_api_v1_companies__company_id__meetings_upload_test_dump_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_test_meeting_dump_api_v1_companies__company_id__meetings_upload_test_dump_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingTestUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Meetings",
        "description": "List all meetings for a company, ordered by date descending.",
        "operationId": "list_meetings_api_v1_companies__company_id__meetings_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/meetings/{meeting_id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting",
        "description": "Get a specific meeting.",
        "operationId": "get_meeting_api_v1_companies__company_id__meetings__meeting_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_meeting_docx_api_v1_companies__company_id__meetings_upload_docx_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_meeting_docx_api_v1_companies__company_id__meetings_upload_docx_post"
      },
      "Body_upload_test_meeting_dump_api_v1_companies__company_id__meetings_upload_test_dump_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_test_meeting_dump_api_v1_companies__company_id__meetings_upload_test_dump_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingList": {
        "properties": {
          "meetings": {
            "items": {
              "$ref": "#/components/schemas/MeetingRead"
            },
            "type": "array",
            "title": "Meetings",
            "description": "List of meetings"
          }
        },
        "type": "object",
        "required": [
          "meetings"
        ],
        "title": "MeetingList"
      },
      "MeetingRead": {
        "properties": {
          "meeting_date": {
            "type": "string",
            "format": "date",
            "title": "Meeting Date",
            "description": "Date when the meeting took place",
            "examples": [
              "2024-03-15"
            ]
          },
          "raw_notes": {
            "type": "string",
            "title": "Raw Notes",
            "description": "Raw meeting notes text content",
            "examples": [
              "Meeting with CEO to discuss Q1 roadmap..."
            ]
          },
          "docx_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docx Filename",
            "description": "Original DOCX filename",
            "examples": [
              "meeting_2024_03_15.docx"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique meeting identifier"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id",
            "description": "Company this meeting belongs to"
          },
          "ai_summary": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary",
            "description": "AI-generated summary of the meeting (JSON array of strings)"
          },
          "decisions": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Decisions",
            "description": "Key decisions made during the meeting (JSON array)"
          },
          "risks": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Risks",
            "description": "Identified risks and concerns (JSON array)"
          },
          "gap_assessment": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Gap Assessment",
            "description": "Assessment of gaps between expectations and reality"
          },
          "alignment_points": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Alignment Points",
            "description": "Points of alignment between VC and founder (JSON array)"
          },
          "gantt_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gantt Status",
            "description": "Status of Gantt chart alignment",
            "examples": [
              "on-track",
              "delayed",
              "ahead"
            ]
          },
          "gantt_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gantt Notes",
            "description": "Notes about Gantt chart progress"
          },
          "commitments": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Commitments",
            "description": "Commitments extracted from meeting (deprecated - use commitments endpoint)"
          },
          "vc_recommendations": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Vc Recommendations",
            "description": "VC's recommendations and advice (JSON array)"
          },
          "initiatives": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Initiatives",
            "description": "New initiatives discussed (JSON array)"
          },
          "financials_mentioned": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Financials Mentioned",
            "description": "Financial metrics and discussions (JSON object)"
          },
          "sentiment": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment",
            "description": "Overall sentiment score (1-5, where 1=very negative, 5=very positive)"
          },
          "sentiment_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Reason",
            "description": "Explanation of the sentiment score"
          },
          "parsed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At",
            "description": "Timestamp when AI parsing was completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Record creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "meeting_date",
          "raw_notes",
          "id",
          "company_id",
          "created_at"
        ],
        "title": "MeetingRead"
      },
      "MeetingTestUploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the upload was successful",
            "examples": [
              true
            ]
          },
          "meetings_processed": {
            "type": "integer",
            "title": "Meetings Processed",
            "description": "Number of meetings processed from the dump file",
            "examples": [
              3
            ]
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results",
            "description": "Results for each processed meeting"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meetings_processed",
          "results"
        ],
        "title": "MeetingTestUploadResponse"
      },
      "MeetingUploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the upload was successful",
            "examples": [
              true
            ]
          },
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id",
            "description": "UUID of the created meeting",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "meeting_date": {
            "type": "string",
            "title": "Meeting Date",
            "description": "Date of the meeting in ISO format",
            "examples": [
              "2024-03-15"
            ]
          },
          "commitments_count": {
            "type": "integer",
            "title": "Commitments Count",
            "description": "Number of commitments extracted from the meeting",
            "examples": [
              5
            ]
          }
        },
        "type": "object",
        "required": [
          "success",
          "meeting_id",
          "meeting_date",
          "commitments_count"
        ],
        "title": "MeetingUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}